Refactor storage service and logging: update import paths and remove unused Post model definition; adjust logging configuration
