+++ backend/src/api/services/gemini_service.py
@@ -1,24 +1,45 @@
 import requests
+import time
 from ..core.config import settings
 from ..core.logging import logger
 
+MAX_RETRIES = 3
+INITIAL_BACKOFF = 1  # seconds
+
 def generate_content(prompt: str) -> str:
-    try:
-        url = "https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash-latest:generateContent"
-        params = {"key": settings.GEMINI_API_KEY}
-        json_data = {
-            "contents": [{"parts": [{"text": f"Generate an engaging Twitter post for an e-commerce brand: {prompt}"}]}],
-            "generationConfig": {"maxOutputTokens": 280},
-        }
-        response = requests.post(url, params=params, json=json_data, timeout=5)
-        response.raise_for_status()
-        generated_text = response.json()["candidates"][0]["content"]["parts"][0]["text"]
-        logger.info(f"Generated content for prompt: {prompt}")
-        return generated_text
-    except requests.exceptions.RequestException as e:
-        logger.error(f"Gemini API error: {str(e)}")
-        return "Failed to generate content due to API error"
-    except KeyError as e:
-        logger.error(f"Unexpected response format from Gemini API: {str(e)}")
-        return "Failed to parse API response"
+    """
+    Generates engaging social media content using the Gemini API.
+
+    Args:
+        prompt (str): The user's prompt or topic for content generation.
+
+    Returns:
+        str: The generated content suitable for a Twitter post, or an error message if generation fails.
+    """
+    for attempt in range(MAX_RETRIES):
+        try:
+            url = "https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash-latest:generateContent"
+            params = {"key": settings.GEMINI_API_KEY}
+            json_data = {
+                "contents": [{"parts": [{"text": f"Generate an engaging Twitter post for an e-commerce brand: {prompt}"}]}],
+                "generationConfig": {"maxOutputTokens": 280},
+            }
+            response = requests.post(url, params=params, json=json_data, timeout=10) # Increased timeout
+            response.raise_for_status()
+            generated_text = response.json()["candidates"][0]["content"]["parts"][0]["text"]
+            logger.info(f"Generated content for prompt: {prompt}")
+            return generated_text
+        except requests.exceptions.RequestException as e:
+            logger.warning(f"Gemini API request failed (attempt {attempt + 1}/{MAX_RETRIES}): {str(e)}")
+            if attempt < MAX_RETRIES - 1:
+                sleep_time = INITIAL_BACKOFF * (2 ** attempt)
+                logger.info(f"Retrying in {sleep_time} seconds...")
+                time.sleep(sleep_time)
+            else:
+                logger.error(f"Gemini API request failed after {MAX_RETRIES} attempts: {str(e)}")
+                return "Failed to generate content due to API error"
+        except KeyError as e:
+            logger.error(f"Unexpected response format from Gemini API: {str(e)}")
+            return "Failed to parse API response"
+    return "Failed to generate content after multiple retries"
 ```
